Custom sbt plugin to fetch dependencies
Using `--sbt-dir` to setup `sbt-dependency-graph` has rather serious
limitations that prevent us from initializing sbt and running the
plugin.
Instead, we add the plugin info with `addPluginSbtFile` and run sbt in
the root path of the project.
`dependencyList::toFile` is problematic for multi-project setups because
it prevents us from writing to the same file for subprojects. Using
`--force` will only override the write and save the dependency list from
the last project analyzed, whichever that might be.

I forked the original project and added `--append` to the original
`toFile` subtask to support bulk collection of dependencies.
We "could" instead look at the list of projects and run `dependencyList`
on each of them but that requires cumbersome and fragile sbt output
parsing.

Note1: We are currently relying on `sbt-dependency-graph` released
quickly to `com.michaelpollmeier` organization. Will switch to
`io.shiftleft` once it lands on Sonatype.
Note2: I didn't try to push a patch upstream as elements of
`sbt-dependency-graph` have been integrated in sbt 1.4.0 and it's
unlikely any work will continue on the plugin itself. We might still
want to provide a similar patch to sbt itself but it will be a longer
path.