Use SBT version when modifying plugins file
A convenient `addPluginSbtFile` is only supported since 1.2.0.
There are plenty of projects still using SBT earlier than that.
To support such projects we
a) detect the correct version
b) modify the plugins' file for SBT < 1.2.0 **in place**
c) run the task
d) cleanup

A possible improvement would be to read from `sbt sbtVersion` but
that feels like an overkill for now. It would also involve parsing the
always evolving sbt output format.