Fix parsing of go.mod file
This fixes an issue with parsing go.mod files if they include package
references that end in `go`. For example, `github.com/aws/aws-sdk-go`.
The matching against the go.mod file is overly aggressive for keywords
like `go` in the file. This aggressive matching results in packages
being dropped from  the scan.

This change ensures that the start of the line is used
in the matching for `go` and `module` keywords. I believe this is safe as
the go.mod files are uniform in construction (at least around these two
areas).

It may be worth making other matches for replace and require more
precise, but that is not part of the scope of this change.