7977: Bring back support for complex sbt projects
SBT project doesn't have to be defined at the root level of project.
Instead it can be defined in the project definition directory. This is
not recommended but still possible, especially for 0.13.x projects.

This change mostly brings back changes that were made in
https://github.com/ShiftLeftSecurity/cdxgen/commit/a11d7fe1387d94831b32c5f69fc281796b9d71e5
but were override when syncing with upstream in
https://github.com/ShiftLeftSecurity/cdxgen/pull/14