Workaround sbt scenario when plugin is present
`sbt-dependency-graph` may already be present in the users' plugins
file. Since we are using our own version of it, and require the most
recent one, we have to ensure that there are no conflicts.
Unfortunately there is no version resolution for plugins in SBT, and it
always picks the first one that is defined (which is usually the bad
one). Rather than pre-pending the plugin, we just make sure there are
no conflicts.