Introduce subproject parameter
When dealing with large and complex monorepos, even non-recursive execution
of cdxgen may unnecessairly take a really long time.
This change introduces an additional parameter, `--subproject`, which
will attempt to generate BOM only for a specific subproject of the
monorepo. Currently, only SBT projects will make use of this parameter.