Don't rely on custom sbt-dependency-graph plugin
Instead of relying on custom sbt-dependency-graph we call `dependencyList` for each project individually.
Then we don't need custom `--append` flag of `dependendencyList`.

Yes, it means shifting some complexity from sbt plugin to cdxgen but the challange with previous solution
was to how instruct SBT to use the custom plugin. I couldn't find a way to do it reliably

One the upside, we no longer need to modify `plugins.sbt` file. Previously we were trying to comment out
import of the plugin but it was using a lot of assumptions and was very fragile

A few more test samples for SBT have been added. cdxgen on master doesn't produce any output for
`sbt1.3-plugin` and `sbt1.6-plugin` confirming that this change fixes some issues people encountered
with SBT builds