on windows look for the bat versions
https://github.com/AppThreat/cdxgen/issues/157

When the code searches for gardlew, on windows it never looks for gradlew.bat, so fails to run it.

per the request, the same change was made for maven, but i have no knowledge of how maven is setup, so this will need to be verified by someone who can, or i'll remove it as i only need gradlew working atm.