Add dependency tree for Rust projects
Fill the dependency tree for Rust projects if a Cargo.lock file is
available. The information to construct the tree is available in the
lock file, so the dependency tree construction is simply a transformation.

There does not seem to be a specification for Cargo.lock files, so the
implmentation makes assumptions of its structure based on available
examples. The assumptions should hold reasonably well because the files
should always be generated by Cargo.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>