Handle Cargo.lock files without footer
The pattern to find package sections in Cargo.lock files does not match
the last package if there is no metadata section at the end of the file.
Fix the pattern to account for a package section followed by EOF.

Multiple Cargo.lock files in a project may produce a dependency list
with duplicate entries, which is not compatible with the CycloneDX
specificaiton. Remove duplicate dependency entries.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>