Handle undefined packages in Rust dependency trees
In case a package is listed as a dependency in a Cargo.lock-file but is
not otherwise defined as a package, it is treated as an error. Continue
gracefully with an error message instead of failing when such a
dependency is encountered.

Add a test to make sure the function does not fail, with a description
of how it has been decided to work for now.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>