Make sure last package of Cargo.lock is included
In certain situations, the components list will miss the last package in
a Cargo.lock-file. A package is only added to the component list when
the Cargo.lock's next package is found. But when end-of-file is
encountered, the last package may be missed.

Fix the issue by adding the current pacakge to the list when the
end-of-file has been reached. Also add a very small test case to make
sure the implementation will include the last package.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>