Remove duplicate BOM-refs in Rust dependency list
If multiple different Cargo.lock-files lists a package of the same
version, there may occur duplicate BOM-references in the dependencies
list.

Merge the dependency lists of packages with identical BOM-refs.

The function `mergeDependencies` cannot be used to merge the lists,
because Rust has no "parent component" which the
`mergeDependencies`-function requires. Therefore it will be implemented
similarly, but not identically. The `mergeDependencies`-function can be
used when the `metadata.component` object of the BOM is created.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>