Add fallback metadata.component for Rust projects
A metadata.component object is needed to make use of the function
`mergeDependencies`. Create a reasonable metadata.component object from
the information available and remove the duplicate implementation of
`mergeDependencies`. Solving issue #939 is meant to find the information
from Cargo.toml-files instead.

Fix a test failing on Windows. The test always replaced `\n` with `\r\n`
causing the test on Windows to produce invalid newlines. The fix will
replace both Windows and Linux newlines with Windows newlines for the
test data.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>