Add debug print to clarify chosen parent component
In order to help the user identify possible erronous parent component
from projects with many Cargo.toml-files, add debug prints to clarify
which one was chosen from which file, and in case any workspace sections
are encountered.

In the case of workspaces, the top-level Cargo.toml file may not contain
a package section describing the project. The top-level Cargo.toml in
rust-lang/rust is one such example. For such situations, describe to the
user that the parent component may be incorrect.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>