Make parseCargoTomlData order independent
Make parseCargoTomlData independent of the order of [package] and
[dependencies] sections. Cargo documentation specifies a convention to
place [package] first, but it is not enforced. Because the parent
component is currently decided based on the ordering of the component
list, it is reasonable to make an effort to make sure the one described
in [package] is chosen.

The API of the parseCargoTomlData is kept intact. However, it would be a
nice future improvement to make the function be explicit about the
seection under which the packages appeared in the file.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>