Suggest solutions for review comments (#956)
* Add debug print to clarify chosen parent component

In order to help the user identify possible erronous parent component
from projects with many Cargo.toml-files, add debug prints to clarify
which one was chosen from which file, and in case any workspace sections
are encountered.

In the case of workspaces, the top-level Cargo.toml file may not contain
a package section describing the project. The top-level Cargo.toml in
rust-lang/rust is one such example. For such situations, describe to the
user that the parent component may be incorrect.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Make parseCargoTomlData order independent

Make parseCargoTomlData independent of the order of [package] and
[dependencies] sections. Cargo documentation specifies a convention to
place [package] first, but it is not enforced. Because the parent
component is currently decided based on the ordering of the component
list, it is reasonable to make an effort to make sure the one described
in [package] is chosen.

The API of the parseCargoTomlData is kept intact. However, it would be a
nice future improvement to make the function be explicit about the
seection under which the packages appeared in the file.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

---------

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>