Return error response in submitBom
When an error occurs, there is no possibility for the calling code to know the details.

In my case, I'd like to be able to log error response body, in particular to be able to know which validation is KO when uploading a SBOM to DependencyTrack (that recently enabled JsonSchema validation on upload).

Signed-off-by: Maxime Robert <robert.maxime@gmail.com>
Signed-off-by: Maxime Robert <maxime.robert@smile.fr>