Using pnpm on Windows binaries as well (#2267)
The issue with not being able to use pnpm, was that pnpm links dependencies to its central store,
which on Windows is done with 'junctions' -- which are full paths.
This means that the packages binary doesn't work, because it is linking to non-existing files!
Hoisting the packages creates a flat node_modules folder without any links.

Also, the package 'encoding' can't be optional on MacOS and Windows.

Signed-off-by: Roland Asmann <roland.asmann@gmail.com>